requirements.txt
colorama==0.4.4
Pillow==8.3.1
